requirements.txt
accelerate==1.8.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appscript==1.3.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
docx2pdf==0.1.8
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
filelock==3.18.0
flasgger==0.9.7.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
ftfy==6.3.1
grpcio==1.67.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
json_repair==0.47.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
milvus-lite==2.5.0
mineru==2.0.6
mistune==3.1.3
modelscope==1.27.1
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.90.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pdfminer.six==20250506
pdftext==0.6.3
pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
pymilvus==2.5.11
pyparsing==3.2.3
pypdf==5.6.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rapid-table==1.0.5
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
robust-downloader==0.0.2
rpds-py==0.25.1
s3transfer==0.13.0
safetensors==0.5.3
scipy==1.16.0
seaborn==0.13.2
shapely==2.1.1
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
stringzilla==3.12.5
sympy==1.14.0
thop==0.1.1.post2209072238
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
